+++ src/main/java/de/samply/project/ProjectType.java
@@ -2,6 +2,7 @@ package de.samply.project;
 
 public enum ProjectType {
     EXPORT,
+    SAMPLES, // Interacts with Negotiator
     DATASHIELD,
     RESEARCH_ENVIRONMENT
 }